Added code to flow

+++ app/Http/Controllers/AuthController.php
@@ -9,20 +9,22 @@ use Illuminate\Support\Facades\Mail;
 use App\Mail\ResetPasswordEmail;
 use App\Mail\ConfirmEmail;
 use App\Mail\WelcomeEmail;
+
 class AuthController extends Controller
 {
-    public function register(Request $request){
+    public function register(Request $request)
+    {
         //Validate data
         //Check if email or username already exists
         $user = User::where('email', $request->email)->orWhere('username', $request->username)->first();
-        if($user){
+        if ($user) {
             return response()->json([
                 'status' => 'failed',
                 'message' => 'Email or username already exists'
             ]);
         }
         //phone number is valid
-        if(strlen($request->phone_number) != 11){
+        if (strlen($request->phone_number) != 11) {
             return response()->json([
                 'status' => 'failed',
                 'message' => 'Phone number is invalid'
@@ -50,19 +52,20 @@ class AuthController extends Controller
             'message' => 'User created successfully'
         ]);
     }
-    public function login(Request $request){
+    public function login(Request $request)
+    {
         $user = $request->user;
         $password = $request->password;
         //Check if user exists
         $user = User::where('email', $user)->orWhere('username', $user)->first();
-        if(!$user){
+        if (!$user) {
             return response()->json([
                 'status' => 'failed',
                 'message' => 'Invalid Data'
             ]);
         }
         //Check if password is correct
-        if(!password_verify($password, $user->password)){
+        if (!password_verify($password, $user->password)) {
             return response()->json([
                 'status' => 'failed',
                 'message' => 'Invalid Data'
@@ -80,17 +83,19 @@ class AuthController extends Controller
         if (User::where('email', $email)->exists()) {
             ///Create password reset token
             $token = bin2hex(random_bytes(64));
+            $code = rand(100000, 999999);
             ///Save token to password_reset_tokens table & remove old tokens
             DB::table('password_reset_tokens')->where('email', $email)->delete();
             DB::table('password_reset_tokens')->insert([
                 'email' => $email,
                 'token' => $token,
+                'code' => $code,
                 'created_at' => now()
             ]);
 
             //Send reset password email
             try {
-                Mail::to($email)->send(new ResetPasswordEmail($token));
+                Mail::to($email)->send(new ResetPasswordEmail($token, $code));
             } catch (\Throwable $th) {
                 //throw $th;
             }
@@ -115,11 +120,35 @@ class AuthController extends Controller
             return response()->json(['status' => 'failed', 'message' => 'Invalid token!']);
         }
     }
+    public function verifyresetPasswordCode(Request $request){
+        $code = $request->code;
+        if (DB::table('password_reset_tokens')->where('code', $code)->exists()) {
+            //Check if code is expired [passed 15mins]
+            $created_at = DB::table('password_reset_tokens')->where('code', $code)->first()->created_at;
+            $created_at = strtotime($created_at);
+            $now = strtotime(now());
+            $diff = $now - $created_at;
+            if ($diff > 900) {
+                return response()->json(['status' => 'failed', 'message' => 'Code expired!']);
+            }
+            return response()->json(['status' => 'success']);
+        } else {
+            return response()->json(['status' => 'failed', 'message' => 'Invalid Code!']);
+        }
+    }
     public function resetPassword(Request $request)
     {
-        $token = $request->token; 
+        $token = $request->token;
         $password = password_hash(strip_tags($request->password), PASSWORD_DEFAULT);
         if (DB::table('password_reset_tokens')->where('token', $token)->exists()) {
+            //Check if token is expired [passed 15mins]
+            $created_at = DB::table('password_reset_tokens')->where('token', $token)->first()->created_at;
+            $created_at = strtotime($created_at);
+            $now = strtotime(now());
+            $diff = $now - $created_at;
+            if ($diff > 900) {
+                return response()->json(['status' => 'failed', 'message' => 'Token expired!']);
+            }
             $email = DB::table('password_reset_tokens')->where('token', $token)->first()->email;
             DB::table('password_reset_tokens')->where('token', $token)->delete();
             DB::table('users')->where('email', $email)->update([
@@ -139,20 +168,22 @@ class AuthController extends Controller
             if ($user->email_verified_at == null) {
                 ///Create email confirm token
                 $token = bin2hex(random_bytes(64));
+                $code = rand(100000, 999999);
                 ///Save token to email_confirm_tokens table & remove old tokens
                 DB::table('email_confirmation_tokens')->where('email', $email)->delete();
                 DB::table('email_confirmation_tokens')->insert([
                     'email' => $email,
                     'token' => $token,
+                    'code' => $code,
                     'created_at' => now()
                 ]);
                 //Send confirm email
                 try {
-                    Mail::to($email)->send(new ConfirmEmail($token));
+                    Mail::to($email)->send(new ConfirmEmail($token, $code));
                 } catch (\Throwable $th) {
                     //throw $th;
                 }
-                return response()->json(['status' => 'success']);           
+                return response()->json(['status' => 'success']);
             } else {
                 return response()->json(['status' => 'failed', 'message' => 'Email already verified!']);
             }
@@ -160,8 +191,17 @@ class AuthController extends Controller
         return response()->json(['status' => 'failed', 'message' => 'Invalid email!']);
     }
     //Verify confirm email token
-    public function verifyConfirmEmail($token){
+    public function verifyConfirmEmail($token)
+    {
         if (DB::table('email_confirmation_tokens')->where('token', $token)->exists()) {
+            //Check if token is expired [passed 15mins]
+            $created_at = DB::table('email_confirmation_tokens')->where('token', $token)->first()->created_at;
+            $created_at = strtotime($created_at);
+            $now = strtotime(now());
+            $diff = $now - $created_at;
+            if ($diff > 900) {
+                return response()->json(['status' => 'failed', 'message' => 'Token expired!']);
+            }
             //Confirm email
             $email = DB::table('email_confirmation_tokens')->where('token', $token)->first()->email;
             DB::table('email_confirmation_tokens')->where('token', $token)->delete();
@@ -171,6 +211,29 @@ class AuthController extends Controller
             return response()->json(['status' => 'success']);
         } else {
             return response()->json(['status' => 'failed', 'message' => 'Invalid token!']);
+        }
+    }
+    //Verify confirm email code
+    public function verifyConfirmEmailCode(Request $request)
+    {
+        $code = $request->code;
+        if (DB::table('email_confirmation_tokens')->where('code', $code)->exists()) {
+            //Check if code is expired [passed 15mins]
+            $created_at = DB::table('email_confirmation_tokens')->where('code', $code)->first()->created_at;
+            $created_at = strtotime($created_at);
+            $now = strtotime(now());
+            $diff = $now - $created_at;
+            if ($diff > 900) {
+                return response()->json(['status' => 'failed', 'message' => 'Code expired!']);
+            }
+            $email = DB::table('email_confirmation_tokens')->where('code', $code)->first()->email;
+            DB::table('email_confirmation_tokens')->where('code', $code)->delete();
+            DB::table('users')->where('email', $email)->update([
+                'email_verified_at' => now()
+            ]);
+            return response()->json(['status' => 'success']);
+        } else {
+            return response()->json(['status' => 'failed', 'message' => 'Invalid Code!']);
         }
     }
 }
